Luke: refactored PMS into DMS
